core code updates
git-svn-id: https://svn.code.sf.net/p/hsqldb/svn/base/trunk@4931 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/navigator/RowSetNavigatorDataTable.java
@@ -52,7 +52,7 @@ import org.hsqldb.rowio.RowOutputInterface;
  * Implementation of RowSetNavigator using a table as the data store.
  *
  * @author Fred Toussi (fredt@users dot sourceforge.net)
- * @version 2.2.7
+ * @version 2.2.9
  * @since 1.9.0
  */
 public class RowSetNavigatorDataTable extends RowSetNavigatorData {
@@ -142,7 +142,7 @@ public class RowSetNavigatorDataTable extends RowSetNavigatorData {
         store              = table.getRowStore(session);
         mainIndex          = table.getPrimaryIndex();
         fullIndex          = table.getFullIndex();
-        this.size          = mainIndex.size(session, store);
+        this.size          = (int) mainIndex.size(session, store);
 
         reset();
     }
